dynarec: Factor patching into seperate file
